backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
email-validator>=2.0